eme: no multiple set server certificate on same mediaKeys
