+++ src/core/eme/set_server_certificate.ts
@@ -23,12 +23,15 @@ import {
 import {
   catchError,
   ignoreElements,
+  tap,
 } from "rxjs/operators";
 import { ICustomMediaKeys } from "../../compat";
 import { EncryptedMediaError } from "../../errors";
 import log from "../../log";
 import castToObservable from "../../utils/cast_to_observable";
+import hashBuffer from "../../utils/hash_buffer";
 import tryCatch from "../../utils/rx-try_catch";
+import ServerCertificateHashStore from "./server_certificate_hash_store";
 import { IEMEWarningEvent } from "./types";
 
 /**
@@ -81,8 +84,24 @@ export default function trySettingServerCertificate(
       return EMPTY;
     }
 
+    const formattedServerCertificate: Uint8Array =
+    serverCertificate instanceof Uint8Array ?
+      serverCertificate :
+      new Uint8Array(
+        serverCertificate instanceof ArrayBuffer ? serverCertificate :
+                                                   serverCertificate.buffer);
+    const currentServerCertificateHash = hashBuffer(formattedServerCertificate);
+    const oldServerCertificateHash = ServerCertificateHashStore.get(mediaKeys);
+
+    if (currentServerCertificateHash === oldServerCertificateHash) {
+      log.debug("EME: Server certificate already set on the MediaKeys");
+      return EMPTY;
+    }
+
     log.info("EME: Setting server certificate on the MediaKeys");
     return setServerCertificate(mediaKeys, serverCertificate).pipe(
+      tap(() => { ServerCertificateHashStore.add(mediaKeys,
+                                                 currentServerCertificateHash); }),
       ignoreElements(),
       catchError(error => observableOf({ type: "warning" as const, value: error })));
   });
